+++ Sources/sol-mac/CLIVersion.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-enum CLIVersion {
-    static let version = "1.0.0"
-    static let build = "1"
-}